+++ packages/web/src/views/earn/components/earn-my-positions/EarnMyPositions.tsx
@@ -24,7 +24,7 @@ export interface EarnMyPositionsProps {
   availableStake: boolean;
   connect: () => void;
   moveEarnAdd: () => void;
-  movePoolDetail: (poolId: string, positionId: string) => void;
+  movePoolDetail: (poolId: string, positionId: number) => void;
   moveEarnStake: () => void;
   mobile: boolean;
   divRef: React.RefObject<HTMLDivElement>;
