+++ packages/web/src/views/pool/pool-remove/components/remove-liquidity/remove-liquidity-select-list/remove-liquidity-select-list-item/RemoveLiquiditySelectListItem.styles.ts
@@ -4,7 +4,7 @@ import { fonts } from "@constants/font.constant";
 import mixins from "@styles/mixins";
 import { media } from "@styles/media";
 
-import { inputStyle } from "../../remove-liquidity/RemoveLiquidity.styles";
+import { inputStyle } from "../../RemoveLiquidity.styles";
 
 export const RemoveLiquiditySelectListItemWrapper = styled.li<{
   selected: boolean;
