+++ packages/web/src/views/pool/pool-remove/components/remove-liquidity/remove-liquidity-select-list/remove-liquidity-select-list-item/RemoveLiquiditySelectListItem.tsx
@@ -24,8 +24,8 @@ import {
 
 interface RemoveLiquiditySelectListItemProps {
   position: PoolPositionModel;
-  checkedList: string[];
-  onCheckedItem: (checked: boolean, path: string) => void;
+  checkedList: number[];
+  onCheckedItem: (checked: boolean, id: number) => void;
   disabled?: boolean;
 }
