+++ packages/web/src/views/pool/pool-remove/components/remove-liquidity/remove-liquidity-select-result/RemoveLiquiditySelectResult.styles.ts
