+++ packages/web/src/views/pool/pool-remove/containers/remove-liquidity-container/RemoveLiquidityContainer.tsx
@@ -10,7 +10,7 @@ import { useRemovePositionModal } from "../../hooks/use-remove-position-modal";
 const RemoveLiquidityContainer: React.FC = () => {
   const router = useCustomRouter();
   const { connected } = useWallet();
-  const [checkedList, setCheckedList] = useState<string[]>([]);
+  const [checkedList, setCheckedList] = useState<number[]>([]);
   const [isGetWGNOT, setIsGetWGNOT] = useState(false);
   const poolPath = router.getPoolPath();
   const { positions, loading: isLoadingPositions } = usePositionData({
@@ -45,12 +45,12 @@ const RemoveLiquidityContainer: React.FC = () => {
   }, [unstakedPositions, checkedList]);
 
   const onCheckedItem = useCallback(
-    (isChecked: boolean, path: string) => {
+    (isChecked: boolean, id: number) => {
       if (isChecked) {
-        return setCheckedList((prev: string[]) => [...prev, path]);
+        return setCheckedList((prev: number[]) => [...prev, id]);
       }
-      if (!isChecked && checkedList.includes(path)) {
-        return setCheckedList(checkedList.filter(el => el !== path));
+      if (!isChecked && checkedList.includes(id)) {
+        return setCheckedList(checkedList.filter(el => el !== id));
       }
     },
     [checkedList],
