+++ packages/web/src/views/pool/pool-stake/components/stake-position/select-liquidity/SelectLiquidity.styles.ts
@@ -1,5 +1,5 @@
 import { css, type Theme } from "@emotion/react";
-import { sectionBoxStyle } from "../stake-position/StakePosition.styles";
+import { sectionBoxStyle } from "../StakePosition.styles";
 
 export const wrapper = (theme: Theme) => css`
   ${sectionBoxStyle(theme)};
