+++ packages/web/src/views/pool/pool-stake/components/stake-position/select-liquidity/SelectLiquidity.tsx
@@ -2,14 +2,14 @@ import React from "react";
 
 import { PoolPositionModel } from "@models/position/pool-position-model";
 
-import SelectLiquidityList from "../select-liquidity-list/SelectLiquidityList";
+import SelectLiquidityList from "./select-liquidity-list/SelectLiquidityList";
 
 import { wrapper } from "./SelectLiquidity.styles";
 
 interface SelectLiquidityProps {
   unstakedPositions: PoolPositionModel[];
-  checkedList: string[];
-  onCheckedItem: (checked: boolean, path: string) => void;
+  checkedList: number[];
+  onCheckedItem: (checked: boolean, path: number) => void;
   onCheckedAll: (checked: boolean) => void;
   checkedAll: boolean;
   isHiddenTitle?: boolean;
