+++ packages/web/src/views/pool/pool-stake/components/stake-position/select-liquidity/select-liquidity-list/SelectLiquidityList.stories.tsx
@@ -9,16 +9,16 @@ export default {
 
 const Template: ComponentStory<typeof SelectLiquidityList> = args => {
 
-  const [checkedList, setCheckedList] = useState<string[]>([]);
+  const [checkedList, setCheckedList] = useState<number[]>([]);
   const [checkedAll] = useState(false);
 
   const onCheckedItem = useCallback(
-    (isChecked: boolean, path: string) => {
+    (isChecked: boolean, id: number) => {
       if (isChecked) {
-        return setCheckedList((prev: string[]) => [...prev, path]);
+        return setCheckedList((prev: number[]) => [...prev, id]);
       }
-      if (!isChecked && checkedList.includes(path)) {
-        return setCheckedList(checkedList.filter(el => el !== path));
+      if (!isChecked && checkedList.includes(id)) {
+        return setCheckedList(checkedList.filter(el => el !== id));
       }
     },
     [checkedList],
