+++ packages/web/src/views/pool/pool-stake/components/stake-position/select-liquidity/select-liquidity-list/SelectLiquidityList.styles.ts
@@ -4,7 +4,7 @@ import { fonts } from "@constants/font.constant";
 import mixins from "@styles/mixins";
 import { media } from "@styles/media";
 
-import { inputStyle } from "../stake-position/StakePosition.styles";
+import { inputStyle } from "../../StakePosition.styles";
 
 export const wrapper = (theme: Theme) => css`
   ${inputStyle(theme)};
