+++ packages/web/src/views/pool/pool-stake/components/stake-position/select-liquidity/select-liquidity-list/SelectLiquidityList.tsx
@@ -10,8 +10,8 @@ import { loadingWrapper, wrapper } from "./SelectLiquidityList.styles";
 
 interface SelectLiquidityProps {
   unstakedPositions: PoolPositionModel[];
-  checkedList: string[];
-  onCheckedItem: (checked: boolean, path: string) => void;
+  checkedList: number[];
+  onCheckedItem: (checked: boolean, path: number) => void;
   onCheckedAll: (checked: boolean) => void;
   checkedAll: boolean;
   isEmpty: boolean;
