+++ packages/web/src/views/pool/pool-stake/components/stake-position/select-liquidity/select-liquidity-list/select-lilquidity-list-item/SelectLiquidityListItem.stories.tsx
@@ -14,7 +14,7 @@ const Template: ComponentStory<typeof SelectLiquidityListItem> = args => {
     <SelectLiquidityListItem
       {...args}
       onCheckedItem={() => setChecked(prev => !prev)}
-      checkedList={checked ? ["#11111"] : []}
+      checkedList={checked ? [111] : []}
     />
   );
 };
