+++ packages/web/src/views/pool/pool-stake/components/stake-position/select-stake-result/SelectStakeResult.stories.tsx
