+++ packages/web/src/views/pool/pool-stake/containers/stake-position-modal-container/StakePositionModalContainer.tsx
@@ -91,7 +91,7 @@ const StakePositionModalContainer = ({
     if (!address) {
       return null;
     }
-    const lpTokenIds = positions.map(position => position.id);
+    const lpTokenIds = positions.map(position => position.id.toString());
     const tokenA = pooledTokenInfos?.[0];
     const tokenB = pooledTokenInfos?.[1];
     broadcastLoading(
