+++ packages/web/src/views/pool/pool-unstake/hooks/use-unstake-position-modal.tsx
@@ -8,7 +8,7 @@ import UnstakePositionModalContainer from "../containers/unstake-position-modal-
 
 interface Props {
   positions: PoolPositionModel[];
-  selectedIds: string[];
+  selectedIds: number[];
   isGetWGNOT: boolean;
 }
